pyproject.toml
[project]
name = "scoptics-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "weaviate-client (>=4.16.9,<5.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "sqlparse (>=0.5.3,<0.6.0)"
]

[tool.poetry]
packages = [{include = "scoptics_agent", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[tool.pytest.ini_options]
pythonpath = "src"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
